Move BraveAPI related code to api folder
