+++ components/brave_vpn/BUILD.gn
@@ -9,19 +9,16 @@ assert(enable_brave_vpn)
 
 static_library("brave_vpn") {
   sources = [
-    "brave_vpn_api_request.cc",
-    "brave_vpn_api_request.h",
     "brave_vpn_service.cc",
     "brave_vpn_service.h",
     "brave_vpn_service_helper.cc",
     "brave_vpn_service_helper.h",
     "brave_vpn_service_observer.cc",
     "brave_vpn_service_observer.h",
-    "vpn_response_parser.cc",
-    "vpn_response_parser.h",
   ]
 
   deps = [
+    "api",
     "common",
     "mojom",
     "//base",
@@ -107,13 +104,13 @@ source_set("unit_tests") {
 
     if (is_win || is_mac || is_android) {
       sources += [
-        "brave_vpn_unittest.cc",
-        "vpn_response_parser_unittest.cc",
+        "brave_vpn_unittest.cc"
       ]
     }
 
     deps = [
       ":brave_vpn",
+      "api:unit_tests",
       "common",
       "//base",
       "//brave/components/p3a_utils",
