+++ components/brave_vpn/api/brave_vpn_api_request.cc
@@ -1,17 +1,17 @@
 /* Copyright (c) 2022 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_vpn/brave_vpn_api_request.h"
+#include "brave/components/brave_vpn/api/brave_vpn_api_request.h"
 
 #include <utility>
 
 #include "base/debug/dump_without_crashing.h"
 #include "base/json/json_writer.h"
+#include "brave/components/brave_vpn/api/brave_vpn_api_helper.h"
+#include "brave/components/brave_vpn/api/vpn_response_parser.h"
 #include "brave/components/brave_vpn/common/brave_vpn_constants.h"
-#include "brave/components/brave_vpn/brave_vpn_service_helper.h"
-#include "brave/components/brave_vpn/vpn_response_parser.h"
 #include "brave/components/skus/browser/skus_utils.h"
 #include "net/traffic_annotation/network_traffic_annotation.h"
 #include "services/network/public/cpp/shared_url_loader_factory.h"
