+++ components/brave_vpn/api/brave_vpn_api_request.h
@@ -1,10 +1,10 @@
 /* Copyright (c) 2022 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_VPN_BRAVE_VPN_API_REQUEST_H_
-#define BRAVE_COMPONENTS_BRAVE_VPN_BRAVE_VPN_API_REQUEST_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_VPN_API_BRAVE_VPN_API_REQUEST_H_
+#define BRAVE_COMPONENTS_BRAVE_VPN_API_BRAVE_VPN_API_REQUEST_H_
 
 #include <string>
 
@@ -96,4 +96,4 @@ class BraveVpnAPIRequest {
 
 }  // namespace brave_vpn
 
-#endif  // BRAVE_COMPONENTS_BRAVE_VPN_BRAVE_VPN_API_REQUEST_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_VPN_API_BRAVE_VPN_API_REQUEST_H_
