+++ components/brave_vpn/api/vpn_response_parser.cc
@@ -1,9 +1,9 @@
 /* Copyright (c) 2022 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/brave_vpn/vpn_response_parser.h"
+#include "brave/components/brave_vpn/api/vpn_response_parser.h"
 
 #include "base/json/json_reader.h"
 #include "base/logging.h"
