+++ components/brave_vpn/brave_vpn_os_connection_api.cc
@@ -12,13 +12,14 @@
 #include "base/check_is_test.h"
 #include "base/json/json_reader.h"
 #include "base/power_monitor/power_monitor.h"
-#include "brave/components/brave_vpn/brave_vpn_api_request.h"
+#include "brave/components/brave_vpn/api/brave_vpn_api_request.h"
+#include "brave/components/brave_vpn/api/brave_vpn_api_helper.h"
 #include "brave/components/brave_vpn/common/brave_vpn_constants.h"
 #include "brave/components/brave_vpn/common/brave_vpn_data_types.h"
 #include "brave/components/brave_vpn/brave_vpn_service_helper.h"
 #include "brave/components/brave_vpn/common/brave_vpn_utils.h"
 #include "brave/components/brave_vpn/common/pref_names.h"
-#include "brave/components/brave_vpn/vpn_response_parser.h"
+#include "brave/components/brave_vpn/api/vpn_response_parser.h"
 #include "components/prefs/pref_service.h"
 #include "third_party/abseil-cpp/absl/types/optional.h"
